refactored select functionality into new file
